Enhanced category disabling and fixed several bugs
Enforced usage of DisabledCategories when deciding or not to register a
listener, or when deciding to return or not in specific listeners; got
rid of the check in checkAchievement, no longer required.
Fixed stats getting collected for all the achievements within a same
listener, even if some categories are disabled.
Various fixes for /aach list.
Updated configuration files.